requirements.txt
numpy
PyPDF2
pandas
tabula-py
nltk
textblob
Flask
flask-restful